transpose hash table
Instead of [ hash val hash val ... ] store [ hash hash ... val val ...].
This leads to the simplest possible API (int*), and should make
vectorized searches of hashes easy.  It's also kind of novel.

Struggled a bit about how to represent this data structure in C, e.g.

    struct hash { union { unsigned hash; int val; } };

but even that is a kind of lie that only works when hash and val are the
same size.  I'd want to express it a bit more generally as an array of N
hashes followed by an array N vals, even if we say stored unsigned hash
and void *val.